pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agente-de-codigo"
version = "0.1.0"
description = "Multi-agent code review system with SAST analysis"
authors = [
    {name = "Felipe", email = "felipe@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["multi-agent", "code-review", "security", "sast", "static-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    # Core dependencies with pinned versions
    "semgrep==1.78.0",
    "gitleaks==8.18.4",
    "ruff==0.1.15",
    "pydantic>=2.5.0,<3.0.0",
    "click>=8.1.0,<9.0.0",
    "rich>=13.7.0,<14.0.0",
    "typer>=0.9.0,<1.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "jsonschema>=4.20.0,<5.0.0",
    "sarif-om==1.0.4",
    "tree-sitter==0.22.4",
    "tree-sitter-python==0.23.6",
    "tree-sitter-typescript==0.23.2",
    "tree-sitter-javascript==0.23.1",
    "lizard==1.17.10",
    "requests>=2.31.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "packaging>=23.2.0,<24.0.0",
    "asyncio-mqtt==0.16.1",
    "websockets>=12.0,<13.0.0",
]

[project.optional-dependencies]
dev = [
    # Development tools
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "black>=23.12.0,<24.0.0",
    "isort>=5.13.0,<6.0.0",
    "mypy>=1.8.0,<2.0.0",
    "flake8>=7.0.0,<8.0.0",
    "bandit>=1.7.5,<2.0.0",
    "safety>=2.3.0,<3.0.0",
    "pre-commit>=3.6.0,<4.0.0",
]

test = [
    # Testing dependencies
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "pytest-xdist>=3.5.0,<4.0.0",
    "factory-boy>=3.3.0,<4.0.0",
    "faker>=20.1.0,<21.0.0",
]

docs = [
    # Documentation dependencies
    "mkdocs>=1.5.0,<2.0.0",
    "mkdocs-material>=9.4.0,<10.0.0",
    "mkdocstrings[python]>=0.24.0,<1.0.0",
    "mkdocs-gen-files>=0.5.0,<1.0.0",
    "mkdocs-literate-nav>=0.6.0,<1.0.0",
]

security = [
    # Security scanning tools
    "bandit>=1.7.5,<2.0.0",
    "safety>=2.3.0,<3.0.0",
    "pip-audit>=2.6.0,<3.0.0",
    "semgrep==1.78.0",
]

performance = [
    # Performance monitoring
    "memory-profiler>=0.61.0,<1.0.0",
    "py-spy>=0.3.14,<1.0.0",
    "line-profiler>=4.1.0,<5.0.0",
]

[project.urls]
Homepage = "https://github.com/felipe/agente-de-codigo"
Documentation = "https://agente-de-codigo.readthedocs.io"
Repository = "https://github.com/felipe/agente-de-codigo.git"
"Bug Tracker" = "https://github.com/felipe/agente-de-codigo/issues"

[project.scripts]
agente = "scripts.run_agents:main"
agente-security = "agents.security_agent.main:main"
agente-quality = "agents.quality_agent.main:main"
agente-architecture = "agents.architecture_agent.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*", "agents*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.toml", "*.txt"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["agents", "scripts", "shared"]
known_third_party = ["semgrep", "ruff", "tree_sitter"]

# mypy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "semgrep.*",
    "ruff.*",
    "tree_sitter.*",
    "lizard.*",
    "gitleaks.*",
]
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=agents",
    "--cov=scripts",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=80",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "security: marks tests as security-related",
    "quality: marks tests as quality-related",
    "architecture: marks tests as architecture-related",
]

# Coverage configuration
[tool.coverage.run]
source = ["agents", "scripts"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/.*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

# Bandit security configuration
[tool.bandit]
exclude_dirs = ["tests", "test_*", "build", "dist"]
skips = ["B101", "B601"]